installer: bulk commit
Squash merge of the following commits from the dev branch:

5ffb16cd1d32836b14110a7083a4e9bb86e83993 ("image-builder: add ls1046a support")
9b67b2532a3a915dbbecbb8b1ea621af5417bab9 ("rpm-ostree: Added rpm-ostree and its dependent pkgs")
16d923c8e57bf136062ab305108cb0ec068066b2 ("installer: make it runnable as an app")
694a83894bcbc3365057b955ccdf69e4c31670f0 ("installer: do not remove input RPMS")
91904d10b124542c76af0c90d68edf7cc72ab7c0 ("Makefile: fix $(PHOTON_STAGE) target")
c50f9c5e96cc0a0d5072ba0b10d595624a4042b5 ("image builder: separate iso builder from installer")
e5be3552cefe0bf973423d6213084857ae06e236 ("installer: deprecate 'iso_system'")
661da1302e494a25a9dcdeefbb951f9d2e8f79d1 ("installer: move partitioning to execution stage")
17096ae22e906b8d5b511c24e75e1a71d4cf73e1 ("installer: use logging")
f1585eba1a83dba4093ea71977f57e768ae53ab5 ("rpm-ostree: Adding server script as sub-package")
a935ac9148d46d1864d7c29ec0ace20d8948a704 ("installer: always use tdnf")
f42f786030cc0d418ca8f73316efbb2434c58e16 ("installer: do not abuse install_config")
8ed3351105fbbc3c085873f3d4eaec0f0bcfa32f ("image-builder: specify dbpath for rpm initdb")
6889b0191f35b5f1044876eb8f707f3b52f12ba6 ("installer: image file partitioning support")
b291a1d201b2cdecbad31e4bdb8fef414780b3db ("installer: reorganize entry point")
2db539aeb5cc26384be843b5f46f04ff55b27635 ("Minor improvements in image builder scripts")
9273a970614162b6813fc61c0442becc83b79403 ("installer: refactor UI initialization")
3799cb7c2eec543dfcb02897358bb5bcf1dbb1b5 ("installer/builder: get additional_rpms_path from config files")
c74fa56d1e0dbc0f75bbb8f520f4aafa0a35f6ae ("Fix grub screw up in mk-install-iso.sh")
802c5823a74314888c65859e0e93e0d04b4dca6c ("installer: lvm support partition")
3d63ebda3b735b8b702c5a7c6b5fdadd3f6dfe33 ("ostree: Adding entries to kernel param in grub.cfg")
c0c4b3cd1ebec5b715588db48d9a66e201a9b14d ("rpm-ostree: Generalise partitioning logic in ostree installer")
7a7f0c892b5814c66815450ec03003d3bc6a09c7 ("ostree: Moved build dependency from installer to image-builder")
f8fdbcecc00a85ff0f6582d1070260aa126a6f81 ("installer: remove shell scripts")
23a5d5280cca27e6efcc07503765f859b52c903f ("Installer: add imagegenerator postactions")
0ac99e7f12a643b1898afd502158c73d17d598c6 ("installer, image-builder: set of fixes")
c305f0a93024aed50c1cde0269204bcb5634ec57 ("GCE image fixes + extras")

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/8259
Tested-by: gerrit-photon <photon-checkins@vmware.com>

+++ support/image-builder/imagegenerator.py
@@ -11,91 +11,6 @@ import json
 from utils import Utils
 import ovagenerator
 
-def prepLoopDevice(loop_device_path, mount_path):
-    Utils.runshellcommand(
-            "mount -t ext4 {} {}".format(loop_device_path, mount_path))
-    Utils.runshellcommand("mount -o bind /proc {}".format(mount_path + "/proc"))
-    Utils.runshellcommand("mount -o bind /dev {}".format(mount_path + "/dev"))
-    Utils.runshellcommand("mount -o bind /dev/pts {}".format(mount_path + "/dev/pts"))
-    Utils.runshellcommand("mount -o bind /sys {}".format(mount_path + "/sys"))        
-
-def cleanupMountPoints(mount_path):
-    Utils.runshellcommand("umount -l {}".format(mount_path + "/sys"))
-    Utils.runshellcommand("umount -l {}".format(mount_path + "/dev/pts"))
-    Utils.runshellcommand("umount -l {}".format(mount_path + "/dev"))
-    Utils.runshellcommand("umount -l {}".format(mount_path + "/proc"))
-
-    Utils.runshellcommand("sync")
-    Utils.runshellcommand("umount -l {}".format(mount_path))
-
-def installAdditionalRpms(mount_path, additional_rpms_path):
-    os.mkdir(mount_path + "/additional_rpms")
-    Utils.copyallfiles(additional_rpms_path,
-                       mount_path + "/additional_rpms")
-    Utils.runshellcommand(
-        "chroot {} /bin/bash -c 'rpm -i /additional_rpms/*'".format(mount_path))
-    shutil.rmtree(mount_path + "/additional_rpms", ignore_errors=True)
-    shutil.rmtree(additional_rpms_path, ignore_errors=True)
-
-def writefstabandgrub(mount_path, uuidval, partuuidval):
-    os.remove(mount_path + "/etc/fstab")
-    f = open(mount_path + "/etc/fstab", "w")
-    if uuidval != '':
-        f.write("UUID={}    /    ext4    defaults 1 1\n".format(uuidval))
-    else:
-        f.write("PARTUUID={}    /    ext4    defaults 1 1\n".format(partuuidval))
-    f.close()
-    Utils.replaceinfile(mount_path + "/boot/grub/grub.cfg",
-                        "rootpartition=PARTUUID=.*$",
-                        "rootpartition=PARTUUID={}".format(partuuidval))
-
-def generateUuid(loop_device_path):
-    partuuidval = (Utils.runshellcommand(
-        "blkid -s PARTUUID -o value {}".format(loop_device_path))).rstrip('\n')
-    uuidval = (Utils.runshellcommand(
-        "blkid -s UUID -o value {}".format(loop_device_path))).rstrip('\n')
-    if partuuidval == '':
-        sgdiskout = Utils.runshellcommand(
-            "sgdisk -i 2 {} ".format(disk_device))
-        partuuidval = (re.findall(r'Partition unique GUID.*',
-                                  sgdiskout))[0].split(':')[1].strip(' ').lower()
-
-    if partuuidval == '':
-        raise RuntimeError("Cannot generate partuuid")
-
-    return (uuidval, partuuidval)
-
-def customizeImage(config, mount_path):
-    build_scripts_path = os.path.dirname(os.path.abspath(__file__))
-    image_name = config['image_type']
-    if 'additionalfiles' in config:
-        for filetuples in config['additionalfiles']:
-            for src, dest in filetuples.items():
-                if (os.path.isdir(build_scripts_path + '/' +
-                                  image_name + '/' + src)):
-                    shutil.copytree(build_scripts_path + '/' +
-                                    image_name + '/' + src,
-                                    mount_path + dest, True)
-                else:
-                    shutil.copyfile(build_scripts_path + '/' +
-                                    image_name + '/' + src,
-                                    mount_path + dest)
-    if 'postinstallscripts' in config:
-        if not os.path.exists(mount_path + "/tempscripts"):
-            os.mkdir(mount_path + "/tempscripts")
-        for script in config['postinstallscripts']:
-            shutil.copy(build_scripts_path + '/' +
-                        image_name + '/' + script,
-                        mount_path + "/tempscripts")
-        for script in os.listdir(mount_path + "/tempscripts"):
-            print("     ...running script {}".format(script))
-            Utils.runshellcommand(
-                "chroot {} /bin/bash -c '/tempscripts/{}'".format(mount_path, script))
-        shutil.rmtree(mount_path + "/tempscripts", ignore_errors=True)
-    if 'expirepassword' in config and config['expirepassword']:
-        # Do not run 'chroot -R' from outside. It will not find nscd socket.
-        Utils.runshellcommand("chroot {} /bin/bash -c 'chage -d 0 root'".format(mount_path))
-
 def createOutputArtifact(raw_image_path, config, src_root, tools_bin_path):
     photon_release_ver = os.environ['PHOTON_RELEASE_VER']
     photon_build_num = os.environ['PHOTON_BUILD_NUM']
@@ -112,22 +27,23 @@ def createOutputArtifact(raw_image_path, config, src_root, tools_bin_path):
 
     shutil.move(raw_image_path, new_name)
     raw_image = new_name
+    compressed = True
 
     if config['artifacttype'] == 'tgz':
         print("Generating the tar.gz artifact ...")
         outputfile = (img_path + '/photon-' + config['image_type'] +
                       '-' + photon_release_ver + '-' +
                       photon_build_num + '.tar.gz')
-        generateCompressedFile(raw_image, outputfile, "w:gz")
+        compressed = generateCompressedFile(raw_image, outputfile, "w:gz")
     elif config['artifacttype'] == 'xz':
         print("Generating the xz artifact ...")
         outputfile = (img_path + '/photon-' + config['image_type'] +
                       '-' + photon_release_ver + '-' +
                       photon_build_num + '.xz')
-        generateCompressedFile(raw_image, outputfile, "w:xz")
+        compressed = generateCompressedFile(raw_image, outputfile, "w:xz")
     elif 'vhd' in config['artifacttype']:
         relrawpath = os.path.relpath(raw_image, src_root)
-        vhdname = (os.path.dirname(relrawpath) + '/photon-' +
+        vhdname = ('/photon-' +
                    config['image_type'] + '-' + photon_release_ver + '-' +
                    photon_build_num + '.vhd')
         print("Converting raw disk to vhd ...")
@@ -142,12 +58,15 @@ def createOutputArtifact(raw_image_path, config, src_root, tools_bin_path):
         Utils.runshellcommand(
             "docker run -v {}:/mnt:rw anishs/qemu-img convert {} -O "
             "vpc -o subformat=fixed,force_size {}"
-            .format(src_root, '/mnt/' + relrawpath, '/mnt/' + vhdname))
+            .format(src_root, '/mnt/' + relrawpath, '/mnt/' + os.path.dirname(relrawpath) + vhdname))
         if config['artifacttype'] == 'vhd.gz':
             outputfile = (img_path + '/photon-' + config['image_type'] +
                           '-' + photon_release_ver + '-' +
                           photon_build_num + '.vhd.tar.gz')
-            generateCompressedFile(vhdname, outputfile, "w:gz")
+            compressed = generateCompressedFile(img_path + vhdname, outputfile, "w:gz")
+            # remove raw image and call the vhd as raw image
+            os.remove(raw_image)
+            raw_image = img_path + vhdname
     elif config['artifacttype'] == 'ova':
         ovagenerator.create_ova_image(raw_image, tools_bin_path, config)
     elif config['artifacttype'] == 'raw':
@@ -155,90 +74,49 @@ def createOutputArtifact(raw_image_path, config, src_root, tools_bin_path):
     else:
         raise ValueError("Unknown output format")
 
+    if not compressed:
+        print("ERROR: Image compression failed!")
+        # Leave the raw disk around if compression failed
+        return
     if not config['keeprawdisk']:
         os.remove(raw_image)
 
 def generateCompressedFile(inputfile, outputfile, formatstring):
-    if formatstring == "w:xz":
-        in_file = open(inputfile, 'rb')
-        in_data = in_file.read()
-
-        out_file = open(inputfile+".xz", 'wb')
-        out_file.write(xz.compress(in_data))
-        in_file.close()
-        out_file.close()
-    else:
-        tarout = tarfile.open(outputfile, formatstring)
-        tarout.add(inputfile, arcname=os.path.basename(inputfile))
-        tarout.close()
-
-def generateImage(raw_image_path, additional_rpms_path, tools_bin_path, src_root, config):
-    working_directory = os.path.dirname(raw_image_path)
-    mount_path = os.path.splitext(raw_image_path)[0]
-    build_scripts_path = os.path.dirname(os.path.abspath(__file__))
-    root_partition_no = 2
-    if 'boot' in config and config['boot'] == 'dualboot':
-        root_partition_no = 3
-
-    if os.path.exists(mount_path) and os.path.isdir(mount_path):
-        shutil.rmtree(mount_path)
-    os.mkdir(mount_path)
-    disk_device = (Utils.runshellcommand(
-        "losetup --show -f {}".format(raw_image_path))).rstrip('\n')
-    disk_partitions = Utils.runshellcommand("kpartx -as {}".format(disk_device))
-    device_name = disk_device.split('/')[2]
-    if not device_name:
-        raise Exception("Could not create loop device and partition")
-    loop_device_path = "/dev/mapper/{}p{}".format(device_name, root_partition_no)
-    print(loop_device_path)
-
     try:
-        (uuidval, partuuidval) = generateUuid(loop_device_path)
-        # Prep the loop device
-        prepLoopDevice(loop_device_path, mount_path)
-        # Clear the root password if not set explicitly from the config file
-        if config['passwordtext'] == 'PASSWORD':
-            Utils.replaceinfile(mount_path + "/etc/shadow",
-                                'root:.*?:', 'root:*:')
-        # Clear machine-id so it gets regenerated on boot
-        open(mount_path + "/etc/machine-id", "w").close()
-        # Write fstab
-        writefstabandgrub(mount_path, uuidval, partuuidval)
-        if additional_rpms_path and os.path.exists(additional_rpms_path):
-            installAdditionalRpms(mount_path, additional_rpms_path)
-        # Perform additional steps defined in installer config
-        customizeImage(config, mount_path)
+        if formatstring == "w:xz":
+            in_file = open(inputfile, 'rb')
+            in_data = in_file.read()
+
+            out_file = open(inputfile+".xz", 'wb')
+            out_file.write(xz.compress(in_data))
+            in_file.close()
+            out_file.close()
+        else:
+            tarout = tarfile.open(outputfile, formatstring)
+            tarout.add(inputfile, arcname=os.path.basename(inputfile))
+            tarout.close()
     except Exception as e:
         print(e)
-    finally:
-        cleanupMountPoints(mount_path)
-        Utils.runshellcommand("kpartx -d {}".format(disk_device))
-        Utils.runshellcommand("losetup -d {}".format(disk_device))
+        return False
+    return True
 
-        shutil.rmtree(mount_path)
-        createOutputArtifact(raw_image_path, config, src_root, tools_bin_path)
-    
 if __name__ == '__main__':
     parser = ArgumentParser()
 
     parser.add_argument("-r", "--raw-image-path", dest="raw_image_path")
     parser.add_argument("-c", "--config-path", dest="config_path")
-    parser.add_argument("-a", "--additional-rpms-path", dest="additional_rpms_path")
     parser.add_argument("-t", "--tools-bin-path", dest="tools_bin_path")
     parser.add_argument("-s", "--src-root", dest="src_root")
 
     options = parser.parse_args()
-    if config_path:
+    if options.config_path:
         config = Utils.jsonread(options.config_path)
     else:
         raise Exception("No config file defined")
 
-    generateImage(
+    createOutputArtifact(
                 options.raw_image_path,
-                options.working_directory,
-                options.additional_rpms_path,
-                options.tools_bin_path,
+                config,
                 options.src_root,
-                config
+                options.tools_bin_path
                 )
-
